First step towards modularising event_routes.py
 Changes to be committed:
	modified:   app.py
	new file:   routes/event_routes.py
