+++ app.py
@@ -12,12 +12,14 @@ from services.api_helpers import (
 )
 from routes.auth_routes import auth_bp
 from routes.dashboard_routes import dashboard_bp
+from routes.event_routes import event_bp
 
 
 # --- App Setup ---
 app = Flask(__name__)
 app.register_blueprint(auth_bp)
 app.register_blueprint(dashboard_bp)
+app.register_blueprint(event_bp)
 app.secret_key = os.environ.get("SECRET_KEY", "dev-secret-key")
 app.permanent_session_lifetime = timedelta(minutes=30)
 
@@ -137,79 +139,6 @@ def fetch_events_by_range():
         print(f"[ERROR] /events: {e}")
         return api_error("Failed to fetch events")
 
-
-@app.route("/event/<event_id>")
-@login_required
-def get_event_detail(event_id):
-    start = time.time()
-    try:
-        id_token = session["user"]["id_token"]
-        token_time = time.time()
-
-        url = urljoin(EVENTS_API_URL, f"/events/{event_id}")
-        headers = create_auth_header(id_token, "application/json")
-        with httpx.Client(timeout=10.0) as client:
-            res = client.get(url, headers=headers)
-            res.raise_for_status()
-        fetch_time = time.time()
-
-        event_data = res.json()
-        parse_time = time.time()
-
-        print(f"[DEBUG] Event size: {len(str(event_data))} characters")
-        print(f"TIMING: token={token_time-start:.3f}s fetch={fetch_time-token_time:.3f}s parse={parse_time-fetch_time:.3f}s")
-
-        response = jsonify(event_data)
-        end_time = time.time()
-        print(f"[DEBUG] jsonify() complete in {end_time - parse_time:.3f}s")
-
-        return response
-    except httpx.HTTPError as e:
-        print(f"[ERROR] Fetching event {event_id}: {e}")
-        return api_error("Failed to fetch event detail")
-
-
-
-
-@app.route("/event/<event_id>", methods=["PATCH"])
-@login_required
-def patch_event(event_id):
-    try:
-        id_token = session["user"]["id_token"]
-        event_data = request.get_json()
-        update_event(event_id, event_data, id_token)
-        return jsonify({"success": True})
-    except Exception as e:
-        print(f"[ERROR] Patching event {event_id}: {e}")
-        return api_error(str(e))
-
-@app.route("/event/<event_id>", methods=["DELETE"])
-@login_required
-def delete_event_route(event_id):
-    try:
-        id_token = session["user"]["id_token"]
-        delete_event(event_id, id_token)
-        return jsonify({"success": True})
-    except Exception as e:
-        print(f"[ERROR] Deleting event {event_id}: {e}")
-        return api_error(str(e))
-
-@app.route("/members")
-@login_required
-def get_members():
-    try:
-        id_token = session["user"]["id_token"]
-        unit_id = request.args.get("invitee_id") or session["user"].get("unit_id")
-        url = urljoin(MEMBERS_URL, f"/units/{unit_id}/members")
-        headers = create_auth_header(id_token, "application/json")
-        with httpx.Client(timeout=10.0) as client:
-            res = client.get(url, headers=headers)
-            res.raise_for_status()
-            return jsonify(res.json())
-    except httpx.HTTPError as e:
-        print(f"[ERROR] Fetching members: {e}")
-        return jsonify({"results": []}), 500
-
 @app.route("/events", methods=["POST"])
 @login_required
 def create_event():
@@ -228,6 +157,25 @@ def create_event():
     except httpx.HTTPError as e:
         print(f"[ERROR] Creating event: {e}")
         return api_error(str(e))
+
+
+@app.route("/members")
+@login_required
+def get_members():
+    try:
+        id_token = session["user"]["id_token"]
+        unit_id = request.args.get("invitee_id") or session["user"].get("unit_id")
+        url = urljoin(MEMBERS_URL, f"/units/{unit_id}/members")
+        headers = create_auth_header(id_token, "application/json")
+        with httpx.Client(timeout=10.0) as client:
+            res = client.get(url, headers=headers)
+            res.raise_for_status()
+            return jsonify(res.json())
+    except httpx.HTTPError as e:
+        print(f"[ERROR] Fetching members: {e}")
+        return jsonify({"results": []}), 500
+
+
 
 if __name__ == "__main__":
     app.run(debug=False)
